refactor: change metric controllers and services naming convention
