+++ modules/metric/controllers/show.ts
@@ -1,11 +1,11 @@
 import type { Request, Response } from "harpiats";
 
 import ApiResponse from "app/helpers/ApiResponse";
-import { Services } from "../services";
+import { service } from "../services";
 
 export async function show(request: Request, response: Response) {
   try {
-    const data = await Services.show();
+    const data = await service.show();
 
     return ApiResponse.success(response, data);
   } catch (error: any) {
